requirements.txt
requests
xrpl-py

